Fixed to_time bug in pf reconstructions

+++ tal/reconstruct/pf/__init__.py
@@ -66,7 +66,7 @@ def to_time(fourier_comp: np.ndarray,
     f_pulse, sig_idx, delta_t = aux_param
     # Approximate the time stamps to the fft index
     t_eval = t_eval or np.array([0.0])
-    t_eval_it = int(np.round(t_eval / delta_t))
+    t_eval_it = np.round(t_eval / delta_t).astype(int)
     # Return the result in time domain
     return fI_to_I(fourier_comp, f_pulse, sig_idx, t_eval_it, n_threads)
